requirements.txt
Django>=5.2.4
requests>=2.31.0
pycryptodome>=3.19.0
